Update main.lua
Added most of Player2's controller movement system that activates when a controller is detected otherwise it would default to the arrow keys. Some variables still need to be renamed life joystick needs to be changed to joystick2 or a separate one needs to be made called joystick2. Also player3 seems to be missing part of the movement script for the A button but i havent access to my Pc to test it right now. If you are looking to play the game choose an earlier version.